refactor: editor
