--- apps/web/.tool-versions
@@ -1,1 +1,0 @@
-nodejs 18.2.0
