--- apps/web/src/components/Document/index.ts
@@ -1,2 +1,0 @@
-export { default as CreateDocument } from './Create';
-export { default as ContentForm } from './DetailBlock';
