+++ apps/web/src/pages/content/[id].tsx
@@ -1,18 +1,21 @@
 import { FC } from 'react';
 import Head from 'next/head';
 
-import ContentDetail from 'components/ContentDetail';
 import Page from 'components/PageFrameInner';
+import DocumentView from 'components/DocumentView';
+import { DocumentProvider } from 'components/DocumentView/DocumentContext';
 
 const Index: FC = () => {
   return (
     <>
       <Head>
-        <title>Content Types - Wraft</title>
+        <title>Document - Wraft</title>
         <meta name="description" content="a nextjs starter boilerplate" />
       </Head>
       <Page>
-        <ContentDetail />
+        <DocumentProvider mode="view">
+          <DocumentView />
+        </DocumentProvider>
       </Page>
     </>
   );
