+++ packages/editor/src/index.ts
@@ -1,2 +1,3 @@
 export type { NodeJSON } from "prosekit/core";
 export { Editor } from "./components/editor";
+// export { LiveEditor } from "./components/live-editor";
