+++ packages/ui/src/theme/GlobalStyle.tsx
@@ -1,10 +1,10 @@
-import { createGlobalStyle } from '@xstyled/emotion';
+import { createGlobalStyle } from "@xstyled/emotion";
 
 export const GlobalStyle = createGlobalStyle`
   @import url("https://fonts.googleapis.com/css2?family=Inter:wght@400;500;700&display=swap");
 
   // html {
-  //   font-size: 16px;
+  //   font-size: 15px;
   // }
   body {
     color: var(--theme-ui-colors-text);
